refactor(cli): physically move clusters handlers into cli_clusters.py (ARCH-2)
Real cut-and-paste of 17 _clusters_* / _cmd_clusters_analyze handlers into
cli_clusters.py (19 defs incl helpers). cli.py 6787 -> 6231 (578 deletions).
cli.py re-imports + _sync_cli_dependencies extended. Anti-shortcut verified:
real defs, funcs gone from cli.py, no shim, no circular, full suite 3141 green.
(Single-module batch after a 40-func/3-module batch failed to complete its
cli.py deletions.) Codex move; Claude verified.
